==> 01_load_reviews.py <==
# -*- coding: utf-8 -*-
"""Copy of JSON reviews to dataframe

Automatically generated by Colab.

Original file is located at
    https://colab.research.google.com/drive/1LJLir-w8QuCY8EoUBZthvk27_DJEqQwF
"""

import json
import pandas as pd

file_path = '/content/data_2.json'
with open(file_path, 'r') as file:
    data = json.load(file)

reviews_data = []
count = 0
for result in data['results']:
  asin = result['content']["asin"]
  for review in result['content']['reviews']:
    count += 1
    try:
      review_info = {
      'review_id': review['id'],
      'product_id': asin,
      "title": review["title"],
      "author": review["author"],
      "rating": review["rating"],
      "content": review["content"],
      "timestamp": review["timestamp"],
      "profile_id": review["profile_id"],
      "is_verified": review["is_verified"],
      "helpful_count": review.get('helpful_count', 0),
      "product_attributes": review["product_attributes"]
      }
      reviews_data.append(review_info)
    except:
      print(count)
      break

reviews_df = pd.DataFrame(reviews_data)
reviews_df.to_csv('reviews_2.csv', index=False)
reviews_df

len(reviews_df)

df_1 = pd.read_csv('/content/reviews_1.csv')
df_2 = pd.read_csv('/content/reviews_2.csv')
df_3 = pd.read_csv('/content/reviews_3.csv')
df_4 = pd.read_csv('/content/reviews_4.csv')
df_5 = pd.read_csv('/content/reviews_5.csv')

df_2.head()

merged_df = pd.concat([df_1, df_2, df_3, df_4, df_5], ignore_index=True)
merged_df

merged_df.to_csv('merged_data.csv', index=False)
