==> 03_numpy_practice.py <==
# -*- coding: utf-8 -*-
"""Copy of Introduction to Numpy and Practice Exercises

Automatically generated by Colab.

Original file is located at
    https://colab.research.google.com/drive/1sp9B92vA9HgVU54o7oXfdsHZl2R62M09

# Numpy - An Introduction

Numerical Python, in-short Numpy is an open-source software, which has the fundamental package for scientific computing with Python. It provides a high-performance multidimensional array object and tools for working with these arrays. It also includes a wide range of Mathematical functions such as Linear Algebra, Fourier Transforms and random number generation, which can be applied to arrays.

Numpy is generally used in the following fields
1. Machine Learning
2. DataScience
3. Image and signal processing
4. Scientific Computing
5. Quantum Computing

### Why Numpy?
1. Faster Execution - Numpy arrays are optimised for complex mathematical and statistical operations. Operations on Numpy are upto 50 times faster than native python lists.
2. Used with various libraries - It is extensively used with various libraries like Pandas, Scipy, Scikit-Learn etcetera.

### Installation

First, let's install NumPy. If you are running this in Google Colab, you don't need to install it as it's already available.
"""

!pip install numpy

"""Once Numpy is installed, it should be imported as given below to use numpy's packages in programming"""

import numpy as np # np is the alias name for the imported numpy library.

"""## Numpy Datastructure:
The primary data structure in NumPy is the ndarray, which stands for N-dimensional array. Numpy arrays have the following restrictions.
1. All the elements of the array must be of same datatype
2. Once the array is declared we cannot change the total size of the array.
3. The shape of the array must be rectangular, ,meaning each row of the 2-dimensional array must have the same number of columns.

## Creating Arrays

One way to initialise the numpy array is to use python lists.
"""

# Creating a single dimensional array
array = np.array([1,2,3,4,5,6])
array

"""So now that we have declared a single dimensional array. There are other built-in functions that Numpy offers like, declaring arrays with only zeroes and ones. We will see that and will see how to declare a two dimensional array."""

zero_arr = np.zeros([3,3]) # This code will generate a simple 3-dimensional array with only zeros.
zero_arr

"""Likewise arrays with only ones also can be generated. In Numpy arrays, we can even mention different datatypes for x-axis and y-axis."""

ones_arr = np.ones([3,3],dtype=[('x', 'float'), ('y', 'int')])
ones_arr

"""## Indexing and Slicing Arrays:

Elements of the array can be accessed using the indexes and a portion of the arrays can be sliced and assigned as a new array or can be used as is. With numpy arrays we can perform almost all operation that we do with Pandas lists. However, following are few of the significant functionalities that we can try with numpy arrays.

1. Basic slicing
2. Fancy indexing
3. Boolean indexing
4. Multidimensional array indexing

##### 1. Basic Slicing
"""

arr = np.array([1, 2, 3, 4, 5, 6, 7])

print(arr[1:5]) # Array values starting from index 1 to 5 will be sliced and displayed.

"""##### 2.Fancy Indexing
In NumPy, fancy indexing allows us to use an array of indices to access multiple array elements at once. It can perform more advanced and efficient array operations, including conditional filtering, sorting, and so on.
"""

#Let us see an example of simple indexing and fancy indexing.
array1 = np.array([1, 2, 3, 4, 5, 6, 7, 8])

# select a single element
simple_indexing = array1[3]

print("Simple Indexing:",simple_indexing)   # 4

# select multiple elements,
fancy_indexing = array1[[1, 2, 5, 7]] # elements present in the index, 1, 2, 5 & 7 will be retrieved and displayed as an array.

print("Fancy Indexing:",fancy_indexing)   # [2 3 6 8]

"""##### 3. Boolean Indexing
Boolean indexing allows us to filter an array based on a specific condition. It can be achieved using setting a boolean mask. Boolean mask is used to set the condition.
"""

array1 = np.array([12, 24, 16, 21, 32, 29, 7, 15])

# Now let's create a mask that selects all elements of array1 that are greater than 20.
# The result will be stored in the boolean mask variable as "True/False" value

boolean_mask = array1 > 20
print("Boolean Mask:", boolean_mask)
print("Filtered Array:", array1[boolean_mask])

"""Similarly, we can implement whatever condition required and can use the result for further operations.

##### 4. Multidimensional array indexing

Like single dimension array indexing, multi dimensional arrays also can be indexed. Bt with the help of slicing function
"""

#lets first declare a multidimensional array using arange function (This function generates number with the given range and given dimensions)
# Multidimensional array indexing
a = np.reshape(np.arange(30), (5, 6))  # Create a 5x6 array
print("Original Array:\n", a)

b = a[1:4, 0:2]  # Select elements in rows 1-3 and columns 0-1
print("Sliced Array:\n", b)

"""## Perform Mathematical Operations:

Numpy offers a wide range of mathematical operations starting from simple arithmetic operation to complex scientific calculations.

Let us have a look at the basic arithmetic operations like, sum, mean, max,min etc.

The below mentioned operations can be done on two or more arrays as well.
"""

sample_array = np.array([1,2,4,5,7,8,3,5])
np.sum(sample_array) # Prints the sum of all values in the array.

np.mean(sample_array) # Displays the avarage value of all the items present in the array

np.max(sample_array) # Displays the highest value present in the array

np.min(sample_array) # Displays the lowest value present in the array.

np.power(3,2) # Displays the value for 3 to the power of 2

np.remainder(70,7) # Displays the remainder of the two values given

np.mod(70,6) # performs the modulus operation

"""## Reshaping and Transposing Arrays:

We will see how the reshape and transpose functions work.

Reshape function lets us to change the shape of the array. The shape of an array is the number of elements in each dimension.
By reshaping we can add or remove dimensions or change number of elements in each dimensio

Transpose function lets us change the rows as columns and columns as rows.e.
"""

#Reshape the one dimensional array as a two dimensional array
arr = np.array([1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12])

print("Original Array",arr)

newarr = arr.reshape(4, 3)

print("Reshaped array \n", newarr)

print("Array before Transpose \n", newarr)
print("Array after Transpose \n", newarr.T)

"""## Concatenating and Splitting Arrays:

Numpy arrays can be  and stacked. There are two function that allows us to do he above said operations.
Numpy.concatenate - Allows us to add arrays to the axes' of the already existing array.
Numpy.Stack - Allows us to add two arrays together.

Also we have options to split arrays, horizontally and vertically.

* Show examples of splitting arrays into smaller chunks using numpy.split(), numpy.hsplit(), or numpy.vsplit().
"""

# Concatenating arrays
arr1 = np.array([1, 2, 3])
arr2 = np.array([4, 5, 6])
concat_arr = np.concatenate((arr1, arr2))
print("Concatenated Array:", concat_arr)

# appends given values in the same axes.

arr3 = np.array([11,22,33])
arr4 = np.array([44,55,66])

stacked_arr = np.stack((arr3,arr4),axis=1)
stacked_arr
# stacks the new array in a new axis

# splits the given array to three new arrays.
arr12 = np.array([10,20,30,40,50,60,70,80,90,100])
new_array = np.array_split(arr12,3)
new_array

"""## Statistical Analysis with NumPy:
The very base of statistical analysis is the mean, median, variance, standard deviation, correlation and covariance. Numpy offer built-in functions to perform all these operations.

* Teach users how to perform statistical analysis on data using NumPy functions like numpy.mean(), numpy.median(), numpy.var(), numpy.std(), etc.
* Explain how to calculate correlations and covariance using numpy.corrcoef() and numpy.cov().
"""

arr12 = np.array([10,20,30,40,50,60,70,80,90,100])
np.mean(arr12)

np.median(arr12)

np.std(arr12) #Standar Deviation

np.corrcoef(arr12)

"""## Random Number Generation:

Random numbers can be generated using numpy.random function which implements pseudo-random number generators (PRNGs or RNGs, for short) with the ability to draw samples from a variety of probability distributions
* Guide users on generating random numbers and arrays using functions like numpy.random.rand(), numpy.random.randint(), numpy.random.normal(), etc.
* Show how to simulate random sampling and distributions for statistical analysis.
"""

array = np.random.rand(5)  # This generates an array with 5 values consisting of random numbers between 0 and 1
array

array = np.random.randint(5)  # Each time this function is execute, a random integer within 5 will be generated.
array

"""## File I/O Operations:

Numpy arrays can be saved to a file and viceversa. The functions used to save the array to a file and load an array from the file is as follows

1. numpy.savetxt() or numpy.save()
2. numpy.loadtxt() or numpy.load()
"""

x = np.arange(0, 10, 1)
print("x is:")
print(x)

# X is an array
c = np.savetxt('sample.txt', (x))    #Save the array to the text file..
a = open("sample.txt", 'r') # open text file in read mode

print("the file contains:")

print(a.read()) #Displays the array saved in the text file.

c= np.loadtxt('sample.txt') #loads the array present in the text file to a variable.
c

"""## Exercises

The following exercises can be easily done if you have diligently gone through the above given exercises.

1. Create a 1D array of numbers from 0 to 9.
"""

# Answer 1
import numpy as np
oned = np.array([0,1,2,3,4,5,6,7,8,9])

print(oned)

"""2. Create a 2D array (3x3) of all ones."""

# Answer 2
import numpy as np
twod = np.array([[1,1,1],[1,1,1]])

print(twod)

"""3. Create a 2D array (5x5) with values ranging from 1 to 25, then reshape it to (5x5)."""

# Answer 3
import numpy as np
twod = np.arange(1,26).reshape(5, 5)

print(twod.shape)

"""4. Extract the middle row and column from the 5x5 array."""

# Answer 4
import numpy as np
twod = np.arange(1,26).reshape(5, 5)

print(twod.shape)

middle_row = twod[2,: ]
middle_col = twod[:, 2]
print(middle_row)
print(middle_col)

"""5. Perform element-wise addition and multiplication on two 1D arrays of length 5 with random integers between 1 and 10."""

# Answer 5
import numpy as np

arr1 = np.random.randint(1, 10, 5)
arr2 = np.random.randint(1, 10, 5)

sum_array = arr1 + arr2
prod_array = arr1 * arr2

print('Array 1 is: ', arr1)
print('Array 2 is: ', arr2)
print('The sum of the arrays are: ', sum_array)
print('The product of the arrays are: ', prod_array)

"""6. Subtract the mean of the array created from the 1D array of numbers from 0 to 9."""

# Answer 6
import numpy as np
num1 = np.arange(0,10)
mean = np.mean(num1)
sub = num1 - mean

print('The numbers are' , num1)
print('The mean is ', mean)
print('The results when subtracted are ', sub)

"""7. Calculate sum, mean, median, and standard deviation of a 1D array."""

array = np.arange(1,6)

mean = np.mean(array)
median = np.median(array)
stand = np.std(array)
sum = np.sum(array)

print('The results are as follows:')
print('Mean: ' , mean)
print('Median: ', median)
print('Standard Dev: ', stand)
print('Sum: ', sum)

"""8. Generate random numbers using numpy.random.rand() and numpy.random.randint()."""

numint = np.random.randint(1,11,5)
num = np.random.rand(5)

print(numint)
print(num)

"""9. Save and load a NumPy array to/from a text file."""

import numpy as np

array = np.arange(1,10)

np.savetxt('np array',array)
