==> 04_misc_analysis.py <==
# -*- coding: utf-8 -*-
"""Untitled3.ipynb

Automatically generated by Colab.

Original file is located at
    https://colab.research.google.com/drive/17LljPz21mLr9p8_hDJDe2VXO3EO8p6Sm
"""

import numpy as np
import pandas as pd

from google.colab import files
uploaded = files.upload()

df = pd.read_csv('merged_data.csv')
df.head()
df.isnull().sum()
df['rating'].unique()

reviews = df['content'].values

count = []
for review in reviews:
  word_count = len(review.split())
  count.append(word_count)

review_lengths = np.array(count)

min_rating = df['rating'].min()
max_rating = df['rating'].max()

df['normalized_rating'] = (df['rating'] - min_rating) / (max_rating - min_rating)
df.head()

df['rating'] = df['rating'].astype(np.float64)
df.info()

df.to_csv('cleaned_reviews.csv', index = False)

import os
os.listdir()
