==> 05_eda_part1.py <==
# -*- coding: utf-8 -*-
"""P3.S4.1

Automatically generated by Colab.

Original file is located at
    https://colab.research.google.com/drive/1T_yIKBJYb_W1wFDNAjAchTsem7b-bir0
"""

import pandas as pd
df = pd.read_csv('telecom_churn.csv')

print(df.head())

print('Mean=',df['Total day minutes'].mean())
print('Median =',df['Total day minutes'].median())
print('Mode= ',df['Total day minutes'].mode()[0])
print('Variance = ',df['Total day minutes'].var())
print('Standard Deviation = ',df['Total day minutes'].std())

print(df.isnull().sum())

Q1 = df['Total eve minutes'].quantile(0.25)
Q3 = df['Total eve minutes'].quantile(0.75)

IQR = Q3 - Q1

lower = Q1 - 1.5 * IQR
upper = Q3 + 1.5 * IQR
outliers = df[(df['Total eve minutes'] < lower) | (df['Total eve minutes'] > upper)]

print('Outliers = ',outliers)

df.to_csv('cleaned_data.csv', index = False)

from google.colab import files
files.download('cleaned_data.csv')
